FIP: SC2: support generate usb password efuse pattern [1/1]
PD#SWPL-31294

Problem:
need to support generate usb password efuse pattern

Solution:
support generate usb password efuse pattern

Verify:
SC2 AH219

+++ fip/sc2/bin/efuse-gen.sh
@@ -6,7 +6,7 @@ VENDOR_KEYTOOL=${EXEC_BASEDIR}/../binary-tool/vendor-keytool
 #
 # Settings
 #
-VERSION=0.2
+VERSION=0.3
 
 # Check file
 check_file() {
@@ -37,13 +37,25 @@ check_opt_file() {
     fi
 }
 
+check_opt_boolean() {
+    if [ -n "$2" ]; then
+        if [ "$2" != "true" ] && [ "$2" != "false" ]; then
+            echo Error: invalid value $1: \""$2"\"
+            exit 1
+        fi
+    fi
+}
+
 usage() {
     cat << EOF
 Usage: $(basename $0) --help
        $(basename $0) --version
-       $(basename $0) --input base.efuse.bin \\
+       $(basename $0) [--input base.efuse.bin] \\
                       [--device-roothash device_roothash.bin] \\
                       [--dvgk dvgk.bin] \\
+                      [--dvuk dvuk.bin] \\
+                      [--enable-usb-password true] \\
+                      [--enable-dvuk-derive-with-cid true] \\
                       -o pattern.efuse
        $(basename $0) --audio-id audio_id_value \\
                       -o audio_id.efuse
@@ -58,22 +70,31 @@ function generate_efuse_device_pattern() {
     local patt=$(mktemp --tmpdir)
     local wrlock0=$(mktemp --tmpdir)
     local wrlock1=$(mktemp --tmpdir)
-	local hmac=$(mktemp --tmpdir)
+    local hmac=$(mktemp --tmpdir)
+    local license0=$(mktemp --tmpdir)
+    local license1=$(mktemp --tmpdir)
+    local efusebit=$(mktemp --tmpdir)
 
     # Parse args
     while [ $i -lt $# ]; do
         arg="${argv[$i]}"
-				#echo "i=$i argv[$i]=${argv[$i]}"
+        #echo "i=$i argv[$i]=${argv[$i]}"
         i=$((i + 1))
         case "$arg" in
             --input)
                 input="${argv[$i]}" ;;
             -o)
                 output="${argv[$i]}" ;;
-			--dvgk)
+            --dvgk)
                 dvgk="${argv[$i]}" ;;
+            --dvuk)
+                dvuk="${argv[$i]}" ;;
 			--device-roothash)
                 device_roothash="${argv[$i]}" ;;
+            --enable-usb-password)
+                enable_usb_password="${argv[$i]}" ;;
+            --enable-dvuk-derive-with-cid)
+                enable_dvuk_derive_with_cid="${argv[$i]}" ;;
             *)
                 echo "Unknown option $arg"; exit 1
                 ;;
@@ -84,29 +105,48 @@ function generate_efuse_device_pattern() {
     # Verify args
     if [ -z "$output" ]; then echo Error: Missing output file option -o; exit 1; fi
 
-	check_file input "$input"
-	check_size "$input" 4096
+    check_opt_file input 4096 "$input"
+    check_opt_file dvgk 16 "$dvgk"
+    check_opt_file dvuk 16 "$dvuk"
+    check_opt_file device_roothash 32 "$device_roothash"
 
-	check_opt_file dvgk 16 "$dvgk"
-	check_opt_file device_roothash 32 "$device_roothash"
+    check_opt_boolean enable-usb-password "$enable_usb_password"
+    check_opt_boolean enable-dvuk-derive-with-cid "$enable_dvuk_derive_with_cid"
 
     # Generate empty eFUSE pattern data
-    dd if="$input" of=$patt count=4096 bs=1 &> /dev/null
+    if [ -n "$input" ]; then
+        dd if="$input" of=$patt count=4096 bs=1 &> /dev/null
+    else
+        dd if=/dev/zero of=$patt count=4096 bs=1 &> /dev/null
+    fi
 
     # Construct wrlock bits
-    b_1e2="00"
-    b_1e3="00"
-    b_1fc="00"
+    dd if=$patt of=$wrlock0 bs=16 skip=30 count=1 &> /dev/null
+    dd if=$patt of=$wrlock1 bs=16 skip=31 count=1 &> /dev/null
+    b_1e2=$(xxd -ps -s2 -l1 $wrlock0)
+    b_1e3=$(xxd -ps -s3 -l1 $wrlock0)
+    b_1fc=$(xxd -ps -s12 -l1 $wrlock1)
 
     if [ "$dvgk" != "" ]; then
 		${VENDOR_KEYTOOL} gen-mrk-chknum --chipset=SC2 --mrk-file="$dvgk" --mrk-name=DVGK | grep 'Long checknum: ' | \
             grep "Long checknum:" | sed 's/Long checknum: //' | sed 's/ (.*//' | xxd -r -p > $hmac
 
-		dd if="$dvgk" of="$patt" bs=16 seek=226 count=1 \
+        dd if="$dvgk" of="$patt" bs=16 seek=226 count=1 \
             conv=notrunc >& /dev/null
-		dd if="$hmac" of="$patt" bs=16 seek=194 count=1 \
+        dd if="$hmac" of="$patt" bs=16 seek=194 count=1 \
             conv=notrunc >& /dev/null
-		b_1fc="$(printf %02x $(( 0x$b_1fc | 0x04 )))"
+        b_1fc="$(printf %02x $(( 0x$b_1fc | 0x04 )))"
+    fi
+
+    if [ "$dvuk" != "" ]; then
+        ${VENDOR_KEYTOOL} gen-mrk-chknum --chipset=SC2 --mrk-file="$dvuk" --mrk-name=DVUK | grep 'Long checknum: ' | \
+            grep "Long checknum:" | sed 's/Long checknum: //' | sed 's/ (.*//' | xxd -r -p > $hmac
+
+        dd if="$dvuk" of="$patt" bs=16 seek=227 count=1 \
+            conv=notrunc >& /dev/null
+        dd if="$hmac" of="$patt" bs=16 seek=195 count=1 \
+            conv=notrunc >& /dev/null
+        b_1fc="$(printf %02x $(( 0x$b_1fc | 0x08 )))"
     fi
 
     if [ "$device_roothash" != "" ]; then
@@ -118,8 +158,12 @@ function generate_efuse_device_pattern() {
 		b_1e3="$(printf %02x $(( 0x$b_1e3 | 0x07 )))"
     fi
 
-    echo 00 00 $b_1e2 $b_1e3 00 00 00 00 00 00 00 00 00 00 00 00 | xxd -r -p > $wrlock0
-    echo 00 00 00 00 00 00 00 00 00 00 00 00 $b_1fc 00 00 00 | xxd -r -p > $wrlock1
+    echo $b_1e2 | xxd -r -p > $efusebit
+    dd if=$efusebit of=$wrlock0 bs=1 seek=2 count=1 conv=notrunc >& /dev/null
+    echo $b_1e3 | xxd -r -p > $efusebit
+    dd if=$efusebit of=$wrlock0 bs=1 seek=3 count=1 conv=notrunc >& /dev/null
+    echo $b_1fc | xxd -r -p > $efusebit
+    dd if=$efusebit of=$wrlock1 bs=1 seek=12 count=1 conv=notrunc >& /dev/null
 
     filesize=$(wc -c < $wrlock0)
     if [ $filesize -ne 16 ]; then
@@ -135,11 +179,46 @@ function generate_efuse_device_pattern() {
     fi
     dd if=$wrlock1 of=$patt bs=16 seek=31 count=1 conv=notrunc >& /dev/null
 
-	${BASEDIR_TOP}/aml_encrypt_sc2 --efsproc --input $patt --output $output --option=debug
+
+    dd if=$patt of=$license0 bs=16 skip=0 count=1 &> /dev/null
+    b_007=$(xxd -ps -s7 -l1 $license0)
+    if [ "$enable_usb_password" == "true" ]; then
+        b_007="$(printf %02x $(( 0x$b_007 | 0x80 )))"
+    fi
+    echo $b_007 | xxd -r -p > $efusebit
+    dd if=$efusebit of=$license0 bs=1 seek=7 count=1 conv=notrunc >& /dev/null
+
+    filesize=$(wc -c < $license0)
+    if [ $filesize -ne 16 ]; then
+        echo Internal Error -- Invalid write-lock0 pattern length
+        exit 1
+    fi
+    dd if=$license0 of=$patt bs=16 seek=0 count=1 conv=notrunc >& /dev/null
+
+    dd if=$patt of=$license1 bs=16 skip=1 count=1 &> /dev/null
+    b_015=$(xxd -ps -s5 -l1 $license1)
+    if [ "$enable_dvuk_derive_with_cid" == "true" ]; then
+        b_015="$(printf %02x $(( 0x$b_015 | 0x02 )))"
+    fi
+    echo $b_015 | xxd -r -p > $efusebit
+    dd if=$efusebit of=$license1 bs=1 seek=5 count=1 conv=notrunc >& /dev/null
+
+    filesize=$(wc -c < $license1)
+    if [ $filesize -ne 16 ]; then
+        echo Internal Error -- Invalid write-lock0 pattern length
+        exit 1
+    fi
+    dd if=$license1 of=$patt bs=16 seek=1 count=1 conv=notrunc >& /dev/null
+
+    ${BASEDIR_TOP}/aml_encrypt_sc2 --efsproc --input $patt --output $output --option=debug
 
     rm -f $patt
     rm -f $wrlock0
-	rm -f $wrlock1
+    rm -f $wrlock1
+    rm -f $hmac
+    rm -f $license0
+    rm -f $license1
+    rm -f $efusebit
 }
 
 function append_uint32_le() {
